fix(base): decide an inclusive end from the raw input, not the parsed value
The EUMETSAT rule added earlier inspected the parsed datetime, so it could not
tell end="2026-08-13" from end="2026-08-13 00:00" -- both parse to midnight and
both were widened to end of day. That leaves no way to ask for "up to midnight
exactly", and reintroduces the multi-GB over-fetch at precisely the boundary the
fix was written for. An arithmetic window that lands on a day boundary hits it.

goes already solved this by keying off the raw input: a datetime.date is bare, a
string is bare when it carries no time separator, and the digit-anchored T check
avoids matching a month name containing a "t". Hoist end_is_date_only and
expand_bare_date_end into earthlens.base so there is one implementation, and
have goes re-export them.

eumetsat, radar and earthdata now record the raw granularity in
_check_input_dates and expand through the shared helper. earthdata previously
widened unconditionally, with the same over-fetch; radar carried a copy of the
parsed-midnight heuristic. Three readings of an inclusive end become one.

AbstractDataSource carries a False class default so a backend that never records
it cannot raise AttributeError.

The radar test that claimed to cover a date-only end actually passed
2024-06-01T00:00:00, an explicit instant, and asserted the old widening; it is
split into the two cases it was conflating.
