+++ libs/core/src/earthlens/base/__init__.py
@@ -11,6 +11,8 @@ from earthlens.base._dates import (
     CADENCE_ALIASES,
     WHOLE_WINDOW,
     date_windows,
+    end_is_date_only,
+    expand_bare_date_end,
     resolve_cadence,
     split_time,
     to_datetime,
@@ -75,6 +77,8 @@ __all__ = [
     "close_quietly",
     "crop_to_aoi",
     "date_windows",
+    "end_is_date_only",
+    "expand_bare_date_end",
     "estimate_pixel_dims",
     "FluxableLeaf",
     "HttpClient",
