+++ libs/providers/atmosphere/src/earthlens/radar/backend.py
@@ -39,6 +39,8 @@ from earthlens.base import (
     OutputKind,
     RemoteProduct,
     TemporalExtent,
+    end_is_date_only,
+    expand_bare_date_end,
 )
 from earthlens.radar.catalog import Catalog, Station
 
@@ -209,15 +211,14 @@ class Radar(AbstractDataSource):
         Raises:
             ValueError: If `start` parses later than `end`.
         """
+        self._end_is_date_only = end_is_date_only(end)
         return self._whole_window_extent(start, end, fmt=fmt, resolution="raw")
 
     def _window(self) -> tuple[dt.datetime, dt.datetime]:
         """Return the inclusive scan-time window, extending `end` to its day end."""
-        end = self.time.end_date
-        # A date-only end (midnight) would exclude the whole day's volumes;
-        # extend it to the end of that day so a same-day request matches.
-        if end.hour == end.minute == end.second == 0:
-            end = end.replace(hour=23, minute=59, second=59, microsecond=999999)
+        # A date-only end (midnight) would exclude the whole day's volumes; an
+        # end naming a time means that instant and is left alone.
+        end = expand_bare_date_end(self.time.end_date, date_only=self._end_is_date_only)
         return self.time.start_date, end
 
     def _search(self) -> list[RemoteProduct]:
